+++ rs-react-app/src/api/rickAndMorty.ts
@@ -1,20 +1,6 @@
 import { FetchError } from '../error/FetchError.ts';
 
-export interface Character {
-  id: number;
-  name: string;
-  status: string;
-  species: number;
-  image: string;
-  gender: string;
-}
-
-export interface ApiResponse {
-  info: {
-    pages: number;
-  };
-  results: Character[];
-}
+import type { ApiResponse } from '../types/rickAndMorty.ts';
 
 export async function fetchAll(
   name: string,
