+++ rs-react-app/src/index.css
@@ -29,24 +29,6 @@ h1 {
   line-height: 1.1;
 }
 
-button {
-  border-radius: 8px;
-  border: 1px solid transparent;
-  padding: 0.6em 1.2em;
-  font-size: 1em;
-  font-weight: 500;
-  font-family: inherit;
-  background-color: hsl(0, 0%, 10%);
-  cursor: pointer;
-  transition: border-color 0.25s;
-}
-button:hover {
-  border-color: hsl(235, 100%, 70%);
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
 
 @media (prefers-color-scheme: light) {
   :root {
@@ -55,8 +37,5 @@ button:focus-visible {
   }
   a:hover {
     color: hsl(235, 100%, 72%);
-  }
-  button {
-    background-color:  hsl(0, 0%, 98%);
   }
 }
