+++ rs-react-app/src/types/rickAndMorty.ts
@@ -1,0 +1,15 @@
+export interface Character {
+  id: number;
+  name: string;
+  status: string;
+  species: number;
+  image: string;
+  gender: string;
+}
+
+export interface ApiResponse {
+  info: {
+    pages: number;
+  };
+  results: Character[];
+}
